pyproject.toml
[tool.poetry]
name = "quubert"
version = "0.1.0"
description = "Bot to help manage the Quub Discord server."
authors = ["Daniil Rose <daniil.rose@member.fsf.org>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = ">= 3.9, < 3.11"
hikari = {extras = ["speedups"], version = "^2.0.0-alpha.104"}
uvloop = "^0.16.0"
hikari-lightbulb = "^2.1.0"
gino = "^1.0.1"
alembic = "^1.7.5"
psycopg2-binary = "^2.9.2"
pandas = "^1.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
